refactor(canvas): GenerationCanvas 抽出 useCanvasViewport 解 800 顶死 [audit#10a]
god-component 顶死 800 行硬上限(52 hook)、任何新增即破门。把 pan/zoom+按分类记忆视口
+stage 尺寸观测+视口虚拟化这簇内聚逻辑抽成 useCanvasViewport(沿用已有 hook 抽取范式),
壳组件 800→740 腾出 headroom。纯机械抽取,逻辑逐字一致;refs 入 deps(稳定不多跑)
